+++ ReleaseNotes.md
@@ -1,5 +1,9 @@
 # Release notes for cities-thermal-comfort-modeling (CTCM) processing framework
 
+## 2025/02/25
+1. Buffered areas are now automatically clipped from the MRT results per https://gfw.atlassian.net/browse/CDB-182
+2. Primary files and intermediate files are not impacted by this change.
+
 ## 2025/02/18
 1. Moved the QGIS viewer to the top level of the city folder and renamed to "qgis_viewer.qgs".
 2. Renamed the ".qgis_viewer" folder to ".qgis_data".